Nw algorithm (#7)
* added pip CI actions file

* added reqs files

* initial python commit and a sample unit test

* changed workflow file jupyter dir error and fixe test flake error

* spaces instead of tabs

* no need to build right now

* initial commit for needleman wunsch code

Generic needleman wunsch utility with uniit tests. It is only part way filled in, but the score matrix generator should work.
Not sure if this will pass online builds because of import issue when src not set in path. Might need to include a setup.py

* Update needleman_wunsch.py

a bit of reformatting

* dataclass version update

version update

* reformatting for lint passing

* changed python min version to 3.7 and added git ignore

* added max line length when testing src and test

* mroe code cleanup and changed the the max line to 120 to meet w/ repo's PEP standard

* moved some files around and introduced a minimal setup.py

introduced a very minimal setup.py to get online testing working w/ package building

* fixed setup.py imports

* made paths relative to file instead of relative to test invocation

* added nltk downloader for online build

Co-authored-by: RevGit <rev2021@gmail.com>